+++ news-js/src/components/view/appView.ts
