+++ news-js/src/components/view/news/news.ts
